+++ pkg/search/options.go
@@ -83,7 +83,7 @@ var (
 	SecretName             = newFieldLabel("Secret")
 	SecretPath             = newFieldLabel("Secret Path")
 	ServiceAccountName     = newFieldLabel("Service Account")
-	UpdatedAt              = newFieldLabel("Updated At")
+	Created                = newFieldLabel("Created")
 	VolumeName             = newFieldLabel("Volume Name")
 	VolumeSource           = newFieldLabel("Volume Source")
 	VolumeDestination      = newFieldLabel("Volume Destination")